+++ plugins/src/main/java/com.escodro.android-dynamic.gradle.kts
@@ -9,7 +9,7 @@ plugins {
 val libs: VersionCatalog = extensions.getByType<VersionCatalogsExtension>().named("libs")
 
 android {
-    androidConfig()
+    androidConfig(libs)
 }
 
 dependencies {
