+++ plugins/src/main/java/extension/VersionCatalogLibraries.kt
@@ -1,4 +1,3 @@
-import org.gradle.api.artifacts.ExternalModuleDependencyBundle
 import org.gradle.api.artifacts.MinimalExternalModuleDependency
 import org.gradle.api.artifacts.VersionCatalog
 import org.gradle.api.provider.Provider
@@ -21,13 +20,5 @@ internal val VersionCatalog.composeBom: Provider<MinimalExternalModuleDependency
 internal val VersionCatalog.composeRules: Provider<MinimalExternalModuleDependency>
     get() = getLibrary("composerules")
 
-internal val VersionCatalog.composeBundle: Provider<ExternalModuleDependencyBundle>
-    get() = getBundle("compose")
-
-internal val VersionCatalog.composeTestBundle: Provider<ExternalModuleDependencyBundle>
-    get() = getBundle("composetest")
-
 private fun VersionCatalog.getLibrary(library: String) = findLibrary(library).get()
-
-private fun VersionCatalog.getBundle(bundle: String) = findBundle(bundle).get()
